+++ leaspy/algo/personalize/scipy_minimize.py
@@ -1,4 +1,5 @@
 from pprint import pformat
+import warnings
 
 import torch
 from joblib import Parallel, delayed
@@ -36,6 +37,7 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         }
 
         if self.algo_parameters['use_jacobian']:
+            # <!> this custom params will remain even if falling back to without jacobian when not implemented
             self.minimize_kwargs = {
                 'method': "BFGS",
                 'options': {
@@ -46,7 +48,7 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
             }
 
         # logging function for convergence warnings
-        # (patient_id: str, scipy_optiminize_result_dict) -> None
+        # (patient_id: str, scipy_minize_result_dict) -> None
         if hasattr(settings, 'logger'):
             self.logger = settings.logger
         else:
@@ -133,7 +135,7 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
             Model values minus real values.
         """
 
-        # computation for 1 individual (level dropped after calculuus)
+        # computation for 1 individual (level dropped after computation)
         predicted = model.compute_individual_tensorized(times.unsqueeze(0), individual_parameters).squeeze(0)
 
         return predicted - values
@@ -280,7 +282,7 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
             # result is objective only
             return res['objective'].item()
 
-    def _get_individual_parameters_patient(self, model, times, values, *, patient_id=None):
+    def _get_individual_parameters_patient(self, model, times, values, *, with_jac: bool, patient_id=None):
         """
         Compute the individual parameter by minimizing the objective loss function with scipy solver.
 
@@ -292,6 +294,8 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
             Contains the individual ages corresponding to the given ``values``.
         values : :class:`torch.Tensor` [n_tpts, n_fts]
             Contains the individual true scores corresponding to the given ``times``.
+        with_jac : bool
+            Should we speed-up the minimization by sending exact gradient of optimized function?
         patient_id : str (or None)
             ID of patient (essentially here for logging purposes when no convergence)
 
@@ -302,9 +306,6 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         reconstruction error : :class:`torch.Tensor` [n_tpts, n_features]
             Model values minus real values.
         """
-
-        # optimize by sending exact gradient of optimized function?
-        with_jac = self.algo_parameters.get('use_jacobian', False)
 
         initial_value = self._initialize_parameters(model)
         res = minimize(self.obj,
@@ -318,16 +319,14 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         err_f = self._get_reconstruction_error(model, times, values, individual_params_f)
 
         if not res.success:
-            if not self.algo_parameters['use_jacobian']:
-                res['reconstruction_mae'] = round(err_f.abs().mean().item(),5) # all tpts & fts instead of mean?
-                res['individual_parameters'] = individual_params_f
-                #del res['x']
-
-                self.logger(patient_id, res)
+            # log full results if optimization failed
+            res['reconstruction_mae'] = err_f.abs().mean().item() # all tpts & fts instead of mean?
+            res['individual_parameters'] = individual_params_f
+            self.logger(patient_id, res)
 
         return individual_params_f, err_f
 
-    def _get_individual_parameters_patient_master(self, it, data, model, *, patient_id=None):
+    def _get_individual_parameters_patient_master(self, it, data, model, *, with_jac: bool, patient_id=None):
         """
         Compute individual parameters of all patients given a leaspy model & a leaspy dataset.
 
@@ -335,10 +334,14 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         ----------
         it : int
             The iteration number.
-        model : :class:`.AbstractModel`
-            Model used to compute the group average parameters.
         data : :class:`.Dataset`
             Contains the individual scores.
+        model : :class:`.AbstractModel`
+            Model used to compute the group average parameters.
+        with_jac : bool
+            Should we speed-up the minimization by sending exact gradient of optimized function?
+        patient_id : str (or None)
+            ID of patient (essentially here for logging purposes when no convergence)
 
         Returns
         -------
@@ -348,13 +351,25 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         times = data.get_times_patient(it)  # torch.Tensor[n_tpts]
         values = data.get_values_patient(it)  # torch.Tensor[n_tpts, n_fts]
 
-        individual_params_tensorized, err = self._get_individual_parameters_patient(model, times, values, patient_id=patient_id)
+        individual_params_tensorized, err = self._get_individual_parameters_patient(model, times, values,
+                                                                                    with_jac=with_jac, patient_id=patient_id)
 
         if self.algo_parameters.get('progress_bar', True):
             self._display_progress_bar(it, data.n_individuals, suffix='subjects')
 
         # transformation is needed because of IndividualParameters expectations...
-        return {k: v.item() if k != 'sources' else v.detach().squeeze(0).tolist() for k,v in individual_params_tensorized.items()}
+        return {k: v.item() if k != 'sources' else v.detach().squeeze(0).tolist()
+                for k,v in individual_params_tensorized.items()}
+
+    def is_jacobian_implemented(self, model) -> bool:
+        """Check that the jacobian of model is implemented."""
+        default_individual_params = self._pull_individual_parameters(self._initialize_parameters(model), model)
+        empty_tpts = torch.tensor([[]], dtype=torch.float32)
+        try:
+            model.compute_jacobian_tensorized(empty_tpts, default_individual_params)
+            return True
+        except NotImplementedError:
+            return False
 
     def _get_individual_parameters(self, model, data):
         """
@@ -378,8 +393,15 @@ class ScipyMinimize(AbstractPersonalizeAlgo):
         if self.algo_parameters.get('progress_bar', True):
             self._display_progress_bar(-1, data.n_individuals, suffix='subjects')
 
+        # optimize by sending exact gradient of optimized function?
+        with_jac = self.algo_parameters['use_jacobian']
+        if with_jac and not self.is_jacobian_implemented(model):
+            warnings.warn('In `scipy_minimize` you requested `use_jacobian=True` but it is not implemented in your model'
+                          f'"{model.name}". Falling back to `use_jacobian=False`...')
+            with_jac = False
+
         ind_p_all = Parallel(n_jobs=self.algo_parameters['n_jobs'])(
-            delayed(self._get_individual_parameters_patient_master)(it_pat, data, model, patient_id=id_pat)
+            delayed(self._get_individual_parameters_patient_master)(it_pat, data, model, with_jac=with_jac, patient_id=id_pat)
             for it_pat, id_pat in enumerate(data.indices))
 
         for it_pat, ind_params_pat in enumerate(ind_p_all):